# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Snap is an open telemetry framework designed to simplify the collection,\nprocessing and publishing of system data through a single API.\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "Snap Restful API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "Snap Maintainer",
      "url": "http://snap.io",
      "email": "snap@intel.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/"
    },
    "version": "1.0.0"
  },
  "host": "127.0.0.1:8181",
  "basePath": "/v2",
  "paths": {
...
    "/plugins": {
      "get": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists a list of loaded plugins. An empty list is returned if there is no loaded plugins.",
        "operationId": "getPlugins",
        "responses": {
          "200": {
            "$ref": "#/responses/PluginsResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json",
          "application/x-protobuf",
          "multipart/form-data"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf",
          "multipart/form-data"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "loads a plugin based on input.",
        "operationId": "loadPlugin",
        "parameters": [
          {
            "type": "file",
            "x-go-name": "PluginData",
            "description": "loads a plugin.",
            "name": "pluginData",
            "in": "formData"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginsResponse"
          },
          "201": {
            "$ref": "#/responses/PluginResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/plugins/{ptype}/{pname}/{pversion}": {
      "get": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists a given plugin by its type, name and version. No plugin found error returns if it's not existing.",
        "operationId": "getPlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json",
          "application/x-protobuf",
          "text/plain"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf",
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "unloads a plugin by its type, name and version.Otherwise, an error is returned.",
        "operationId": "unloadPlugin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginsResponse"
          },
          "204": {
            "$ref": "#/responses/PluginResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/plugins/{ptype}/{pname}/{pversion}/config": {
      "get": {
        "description": "Any other type results in error.",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "lists the config of a giving plugin. The allowed plugin types are collector, processor, and publisher.",
        "operationId": "getPluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "updates the config of a giving plugin. A wrong plugin type or non-numeric plugin version results in error.",
        "operationId": "setPluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Config",
            "name": "config",
            "in": "formData"
          },
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "At lease one config key is required for this operation. An error occurs for any bad request.",
        "consumes": [
          "application/json",
          "application/x-protobuf"
        ],
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "deletes the config of a giving plugin. Note that that to be removed config items are a slice of config keys.",
        "operationId": "deletePluginConfigItem",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PName",
            "name": "pname",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "PVersion",
            "name": "pversion",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "collector",
              " processor",
              " publisher"
            ],
            "type": "string",
            "x-go-name": "PType",
            "name": "ptype",
            "in": "path",
            "required": true
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Config",
            "name": "config",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PluginConfigResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ConfigDataNode": {
      "description": "Represents a set of configuration data",
      "type": "object",
      "x-go-package": "github.com/intelsdi-x/snap/core/cdata"
    },
...
    "Plugin": {
      "type": "object",
      "title": "Plugin represents a plugin type definition.",
      "properties": {
        "config_policy": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PolicyTable"
          },
          "x-go-name": "ConfigPolicy"
        },
        "href": {
          "type": "string",
          "x-go-name": "Href"
        },
        "loaded_timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LoadedTimestamp"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "signed": {
          "type": "boolean",
          "x-go-name": "Signed"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        },
        "version": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/mgmt/rest/v2"
    },
    "PolicyTable": {
      "x-go-package": "github.com/intelsdi-x/snap/mgmt/rest/v2",
      "$ref": "#/definitions/RuleTable"
    },
...
    "RuleTable": {
      "type": "object",
      "title": "RuleTable defines a named property characteristics.",
      "required": [
        "name",
        "type",
        "required"
      ],
      "properties": {
        "default": {
          "type": "object",
          "x-go-name": "Default"
        },
        "maximum": {
          "type": "object",
          "x-go-name": "Maximum"
        },
        "minimum": {
          "type": "object",
          "x-go-name": "Minimum"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "required": {
          "type": "boolean",
          "x-go-name": "Required"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/intelsdi-x/snap/control/plugin/cpolicy"
    },
...
  },
  "responses": {
    "ErrorResponse": {
      "description": "Error unsuccessful generic response to a failed API call",
      "schema": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      }
    },
...
    "PluginConfigResponse": {
      "description": "PluginConfigItem represents cdata.ConfigDataNode which implements it's own UnmarshalJSON.",
      "schema": {
        "$ref": "#/definitions/ConfigDataNode"
      }
    },
    "PluginResponse": {
      "description": "PluginResponse represents the response from plugin operations.",
      "schema": {
        "$ref": "#/definitions/Plugin"
      }
    },
    "PluginsResponse": {
      "description": "PluginsResp represents the response from plugin operations.",
      "schema": {
        "type": "object",
        "properties": {
          "plugins": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Plugin"
            },
            "x-go-name": "Plugins"
          }
        }
      }
    },
...
  }
}